batch: Move source advancement out of ownership
The batch package has ownership models, lineage remapping, and stale-source
refresh workflows for keeping batch sources current.

The ownership module still hosts source-advance routines that load
repository buffers and call merge materialization, which couples ownership
back to the merge engine.

This commit addresses that by moving the source-advance routines into
batch.source_advancement and updating production callers to use that
module.

Subsequent commits will update tests and add import-boundary coverage for
the new module.
