+++ src/git_stage_batch/batch/source_refresh.py
@@ -26,12 +26,12 @@ from .match import match_lines
 from .ownership import (
     BatchOwnership,
     ReplacementLineRun,
-    advance_batch_source_for_file_with_provenance,
     detect_stale_batch_source_for_selection,
     merge_batch_ownership,
     translate_hunk_selection_to_batch_ownership,
     translate_lines_to_batch_ownership,
 )
+from .source_advancement import advance_batch_source_for_file_with_provenance
 
 
 @dataclass
